videoSummarizer/requirements.txt
absl-py==0.9.0
appnope==0.1.0
asgiref==3.2.7
astor==0.8.1
attrdict==2.0.1
attrs==19.3.0
audioread==2.1.8
Automat==20.2.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.0
boto==2.49.0
boto3==1.12.39
botocore==1.15.39
botometer==1.4
cachetools==4.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cli-helpers==1.2.1
click==7.1.1
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cvxpy==1.1.0a3
cvxpylayers==0.1.2
cycler==0.10.0
decorator==4.4.1
deepspeech==0.6.1
defusedxml==0.6.0
diffcp==1.0.13
distro==1.4.0
Django==3.0.5
docutils==0.15.2
ecos==2.0.7.post1
efficientnet-pytorch==0.6.3
entrypoints==0.3
ffmpeg-python==0.2.0
filelock==3.0.12
Flask==1.1.1
future==0.18.2
gast==0.3.3
gdown==3.10.2
gensim==3.8.2
google-api-core==1.16.0
google-auth==1.13.1
google-cloud-core==1.3.0
google-cloud-speech==1.3.2
google-cloud-storage==1.27.0
google-pasta==0.2.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
h5py==2.10.0
hyperlink==19.0.0
idna==2.8
ignite==1.1.0
importlib-metadata==1.3.0
incremental==17.5.0
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.2
jellyfish==0.7.2
Jinja2==2.10.3
jmespath==0.9.5
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
librosa==0.7.2
line-profiler==3.0.2
llvmlite==0.31.0
lxml==4.4.2
Markdown==3.2.1
MarkupSafe==1.1.1
mask-rcnn==2.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==8.0.2
munch==2.5.0
mycli==1.21.1
mysqlclient==1.3.13
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
nltk==3.4.5
notebook==6.0.2
numba==0.48.0
numpy==1.18.0
oauthlib==3.1.0
open3d==0.9.0.0
opencv-python==4.2.0.32
osqp==0.6.1
packaging==20.3
pandas==0.25.3
pandocfilters==1.4.2
parsel==1.5.2
parso==0.5.2
pause==0.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pluggy==0.13.1
pretrainedmodels==0.7.4
progress-bar==8
prometheus-client==0.7.1
prompt-toolkit==2.0.10
Protego==0.1.16
protobuf==3.11.3
ptyprocess==0.6.0
punctuator==0.9.5
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
pycparser==2.20
PyDispatcher==2.0.5
pydub==0.23.1
Pygments==2.5.2
PyHamcrest==2.0.2
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
PyQt5==5.14.2
PyQt5-sip==12.7.2
pyrsistent==0.15.6
PySocks==1.7.1
pytest==5.4.1
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytorch-ignite==0.3.0
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.6.0
queuelib==1.5.0
rake-nltk==1.0.4
regex==2020.4.4
requests==2.22.0
requests-oauthlib==1.3.0
resampy==0.2.2
rope==0.16.0
rsa==4.0
s3transfer==0.3.3
scenedetect==0.5.2
scikit-learn==0.22.2.post1
scipy==1.4.1
Scrapy==2.0.0
scs==2.1.2
seaborn==0.9.0
segmentation-models-pytorch==0.1.0
segtok==1.5.9
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.14.0
smart-open==1.11.1
SoundFile==0.10.3.post1
soupsieve==1.9.5
sqlparse==0.3.1
tabula-py==2.1.0
tabulate==0.8.7
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
terminaltables==3.1.0
testpath==0.4.4
Theano==1.0.4
threadpoolctl==2.0.0
torch==1.3.1
torchvision==0.4.2
tornado==6.0.3
tqdm==4.43.0
traitlets==4.3.3
twarc==1.8.1
tweepy==3.8.0
Twisted==19.10.0
urllib3==1.25.7
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yake==0.4.2
zipp==0.6.0
zope.interface==4.7.2
